[FEAT] reorganized import modules
