+++ src/app/modules/homes/body/library/library.module.ts
@@ -7,13 +7,13 @@ import { CategoryComponent } from './category/category.component';
 // import { GraphComponent } from './graph/graph.component';
 import { CatGraphComponent } from './category-graph/category-graph.component';
 // import { SearchModule } from '../search/search.module';
-import { CommonSearchResultDocumentListModule } from '../shared-module/common-search-result-document-list/common-search-result-document-list.module';
-import { CommonSearchBarModule } from '../shared-module/common-search-bar/common-search-bar.module';
+import { CommonSearchResultDocumentListModule } from '../shared-modules/documents/documents.module';
+import { CommonSearchBarModule } from '../shared-modules/search-bar/common-search-bar.module';
 
 @NgModule({
   declarations: [
-    LibraryRootComponent, 
-    CategoryComponent, 
+    LibraryRootComponent,
+    CategoryComponent,
     // GraphComponent,
     CatGraphComponent
   ],
@@ -24,7 +24,7 @@ import { CommonSearchBarModule } from '../shared-module/common-search-bar/common
     CommonSearchResultDocumentListModule,
     CommonSearchBarModule
   ],
-  exports:[
+  exports: [
     LibraryRootComponent,
   ]
 })
